str_add_ch-tune

+++ test.cpp
@@ -239,12 +239,12 @@ void test4()
     setp += "this is, just a simple. but fixed, nonsense test, voila :)";
     }
 
-  printf( "set  = %d\n", str_len( set  ) );
-  printf( "cat  = %d\n", str_len( cat  ) );
-  printf( "setn = %d\n", str_len( setn ) );
-  printf( "catn = %d\n", str_len( catn ) );
-  printf( "sete = %d\n", str_len( sete ) );
-  printf( "setp = %d\n", str_len( setp ) );
+  printf( "set  = %d\n", (int)str_len( set  ) );
+  printf( "cat  = %d\n", (int)str_len( cat  ) );
+  printf( "setn = %d\n", (int)str_len( setn ) );
+  printf( "catn = %d\n", (int)str_len( catn ) );
+  printf( "sete = %d\n", (int)str_len( sete ) );
+  printf( "setp = %d\n", (int)str_len( setp ) );
 
   printf( "--------------------\n" );
 
@@ -391,7 +391,7 @@ void test8()
   v1 = "this is simple test ";
   v1 *= 1024;
 
-  printf( "v1 len: %d\n", str_len( v1 ) );
+  printf( "v1 len: %d\n", (int)str_len( v1 ) );
 
   v2.compact( 1 ); // makes v2 compact, i.e. it will get as much memory as it
                    // needs. otherwise it will get fixed amount of blocks
